+++ cmd/keeman/README.md
@@ -3,6 +3,6 @@ The Keymaker
 This is a simple tool for creating various types of encryption keys.
 
 ```shell
-go get -u github.com/chronicleprotocol/oracle-suite/keeman
+go get -u github.com/chronicleprotocol/oracle-suite/cmd/keeman
 ```
