+++ cmd/keeman/rand/bytes.go
