+++ cmd/keeman/txt/file_test.go
