+++ cmd/keeman/txt/reader.go
